requirements.txt
customtkinter
mysql-connector-python
qrcode[pil]
reportlab
pandas
openpyxl
